In ns_hrtime2nsec(), cast "val" to guint64 to fix Coverity CID 702389 Unintentional integer overflow.
svn path=/trunk/; revision=42846

+++ wiretap/netscaler.c
@@ -476,7 +476,7 @@ static guint64 ns_hrtime2nsec(guint32 tm)
 	{
 	case NSPR_HRTIME_SEC:	return (guint64)val*1000000000;
 	case NSPR_HRTIME_MSEC:	return (guint64)val*1000000;
-	case NSPR_HRTIME_USEC:	return (guint32)val*1000;
+	case NSPR_HRTIME_USEC:	return (guint64)val*1000;
 	case NSPR_HRTIME_NSEC:	return val;
 	}
 	return tm;
